mcp-server/requirements.txt
openai
mcp
pydantic
fastapi
uvicorn
python-dotenv
sqlmodel
psycopg2-binary
pyjwt
cryptography